Migrating to Phoenix 1.3 release structure
* Separating lib in two subfolders, real_world and real_world_web

Phoenix after 1.3 release has changed the structure to be that of two
apps, your app itself and the web part of it. We are now following this
structure as well.

* Renamed all RealWorld.Web to RealWorldWeb

Also following Phoenix 1.3 release recommendation of renaming the
App.Web to AppWeb.

* Deleted all frontend from /priv

Since this is for backend purposes only, all frontend stuff is
not required.

* Deleted model_case
