+++ config/config.exs
@@ -10,10 +10,10 @@ config :real_world,
   ecto_repos: [RealWorld.Repo]
 
 # Configures the endpoint
-config :real_world, RealWorld.Web.Endpoint,
+config :real_world, RealWorldWeb.Endpoint,
   url: [host: "localhost"],
   secret_key_base: "9ueg5YcX8/LKzVUcDrXp5xpYuaBCUfZZAJ3/udC1LCoabotR3O1CJyf/u/6RLJ/N",
-  render_errors: [view: RealWorld.Web.ErrorView, accepts: ~w(json)],
+  render_errors: [view: RealWorldWeb.ErrorView, accepts: ~w(json)],
   pubsub: [name: RealWorld.PubSub,
            adapter: Phoenix.PubSub.PG2]
