+++ config/prod.exs
@@ -3,7 +3,7 @@ use Mix.Config
 # For production, we often load configuration from external
 # sources, such as your system environment. For this reason,
 # you won't find the :http configuration below, but set inside
-# RealWorld.Web.Endpoint.load_from_system_env/1 dynamically.
+# RealWorldWeb.Endpoint.load_from_system_env/1 dynamically.
 # Any dynamic configuration should be moved to such function.
 #
 # Don't forget to configure the url host to something meaningful,
@@ -13,8 +13,8 @@ use Mix.Config
 # containing the digested version of static files. This
 # manifest is generated by the mix phoenix.digest task
 # which you typically run after static files are built.
-config :real_world, RealWorld.Web.Endpoint,
-  on_init: {RealWorld.Web.Endpoint, :load_from_system_env, []},
+config :real_world, RealWorldWeb.Endpoint,
+  on_init: {RealWorldWeb.Endpoint, :load_from_system_env, []},
   url: [host: "example.com", port: 80],
   cache_static_manifest: "priv/static/cache_manifest.json"
 
@@ -26,7 +26,7 @@ config :logger, level: :info
 # To get SSL working, you will need to add the `https` key
 # to the previous section and set your `:url` port to 443:
 #
-#     config :real_world, RealWorld.Web.Endpoint,
+#     config :real_world, RealWorldWeb.Endpoint,
 #       ...
 #       url: [host: "example.com", port: 443],
 #       https: [:inet6,
@@ -41,7 +41,7 @@ config :logger, level: :info
 # We also recommend setting `force_ssl`, ensuring no data is
 # ever sent via http, always redirecting to https:
 #
-#     config :real_world, RealWorld.Web.Endpoint,
+#     config :real_world, RealWorldWeb.Endpoint,
 #       force_ssl: [hsts: true]
 #
 # Check `Plug.SSL` for all available options in `force_ssl`.
@@ -56,7 +56,7 @@ config :logger, level: :info
 # Alternatively, you can configure exactly which server to
 # start per endpoint:
 #
-#     config :real_world, RealWorld.Web.Endpoint, server: true
+#     config :real_world, RealWorldWeb.Endpoint, server: true
 #
 
 # Finally import the config/prod.secret.exs
