+++ lib/real_world/application.ex
@@ -12,7 +12,7 @@ defmodule RealWorld.Application do
       # Start the Ecto repository
       supervisor(RealWorld.Repo, []),
       # Start the endpoint when the application starts
-      supervisor(RealWorld.Web.Endpoint, []),
+      supervisor(RealWorldWeb.Endpoint, []),
       # Start your own worker by calling: RealWorld.Worker.start_link(arg1, arg2, arg3)
       # worker(RealWorld.Worker, [arg1, arg2, arg3]),
     ]
@@ -21,5 +21,12 @@ defmodule RealWorld.Application do
     # for other strategies and supported options
     opts = [strategy: :one_for_one, name: RealWorld.Supervisor]
     Supervisor.start_link(children, opts)
+  end
+
+  # Tell Phoenix to update the endpoint configuration
+  # whenever the application is updated.
+  def config_change(changed, _new, removed) do
+    RealWorldWeb.Endpoint.config_change(changed, removed)
+    :ok
   end
 end
