--- lib/real_world/web/controllers/fallback_controller.ex
@@ -1,20 +1,0 @@
-defmodule RealWorld.Web.FallbackController do
-  @moduledoc """
-  Translates controller action results into valid `Plug.Conn` responses.
-
-  See `Phoenix.Controller.action_fallback/1` for more details.
-  """
-  use RealWorld.Web, :controller
-
-  def call(conn, {:error, %Ecto.Changeset{} = changeset}) do
-    conn
-    |> put_status(:unprocessable_entity)
-    |> render(RealWorld.Web.ChangesetView, "error.json", changeset: changeset)
-  end
-
-  def call(conn, {:error, :not_found}) do
-    conn
-    |> put_status(:not_found)
-    |> render(RealWorld.Web.ErrorView, :"404")
-  end
-end
