--- lib/real_world/web/views/changeset_view.ex
@@ -1,19 +1,0 @@
-defmodule RealWorld.Web.ChangesetView do
-  use RealWorld.Web, :view
-
-  @doc """
-  Traverses and translates changeset errors.
-
-  See `Ecto.Changeset.traverse_errors/2` and
-  `RealWorld.Web.ErrorHelpers.translate_error/1` for more details.
-  """
-  def translate_errors(changeset) do
-    Ecto.Changeset.traverse_errors(changeset, &translate_error/1)
-  end
-
-  def render("error.json", %{changeset: changeset}) do
-    # When encoded, the changeset returns its errors
-    # as a JSON object. So we just pass it forward.
-    %{errors: translate_errors(changeset)}
-  end
-end
