--- lib/real_world/web/views/error_view.ex
@@ -1,17 +1,0 @@
-defmodule RealWorld.Web.ErrorView do
-  use RealWorld.Web, :view
-
-  def render("404.json", _assigns) do
-    %{errors: %{detail: "Page not found"}}
-  end
-
-  def render("500.json", _assigns) do
-    %{errors: %{detail: "Internal server error"}}
-  end
-
-  # In case no render clause matches or no
-  # template is found, let's render it as 500
-  def template_not_found(_template, assigns) do
-    render "500.json", assigns
-  end
-end
