+++ lib/real_world_web/views/error_helpers.ex
@@ -1,0 +1,30 @@
+defmodule RealWorldWeb.ErrorHelpers do
+  @moduledoc """
+  Conveniences for translating and building error messages.
+  """
+
+  @doc """
+  Translates an error message using gettext.
+  """
+  def translate_error({msg, opts}) do
+    # Because error messages were defined within Ecto, we must
+    # call the Gettext module passing our Gettext backend. We
+    # also use the "errors" domain as translations are placed
+    # in the errors.po file.
+    # Ecto will pass the :count keyword if the error message is
+    # meant to be pluralized.
+    # On your own code and templates, depending on whether you
+    # need the message to be pluralized or not, this could be
+    # written simply as:
+    #
+    #     dngettext "errors", "1 file", "%{count} files", count
+    #     dgettext "errors", "is invalid"
+    #
+    case opts[:count] do
+      nil ->
+        Gettext.dgettext(RealWorldWeb.Gettext, "errors", msg, opts)
+      count ->
+        Gettext.dngettext(RealWorldWeb.Gettext, "errors", msg, msg, count, opts)
+    end
+  end
+end
