+++ test/support/channel_case.ex
@@ -1,4 +1,4 @@
-defmodule RealWorld.Web.ChannelCase do
+defmodule RealWorldWeb.ChannelCase do
   @moduledoc """
   This module defines the test case to be used by
   channel tests.
@@ -21,7 +21,7 @@ defmodule RealWorld.Web.ChannelCase do
       use Phoenix.ChannelTest
 
       # The default endpoint for testing
-      @endpoint RealWorld.Web.Endpoint
+      @endpoint RealWorldWeb.Endpoint
     end
   end
