+++ test/support/conn_case.ex
@@ -1,4 +1,4 @@
-defmodule RealWorld.Web.ConnCase do
+defmodule RealWorldWeb.ConnCase do
   @moduledoc """
   This module defines the test case to be used by
   tests that require setting up a connection.
@@ -19,10 +19,10 @@ defmodule RealWorld.Web.ConnCase do
     quote do
       # Import conveniences for testing with connections
       use Phoenix.ConnTest
-      import RealWorld.Web.Router.Helpers
+      import RealWorldWeb.Router.Helpers
 
       # The default endpoint for testing
-      @endpoint RealWorld.Web.Endpoint
+      @endpoint RealWorldWeb.Endpoint
     end
   end
